+++ xbmc/addons/kodi-addon-dev-kit/kodi/addon.api2/definitions.h
